+++ babel.js
@@ -1,6 +1,7 @@
 module.exports = {
     extends: [
         "eslint-config-konnektid",
+        "eslint-config-konnektid/specs/es7",
         "eslint-config-konnektid/rules/babel"
     ]
 };