+++ envs/universal.js
@@ -1,0 +1,7 @@
+module.exports = {
+    env: {
+        "shared-node-browser": true,
+        node: false,
+        browser: false
+    }
+};