+++ envs/webpack.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+    globals: {
+        "require": false,
+        "module": false
+    }
+}