+++ module.js
@@ -1,7 +1,6 @@
 module.exports = {
     extends: [
-        "eslint-config-konnektid",
-        "eslint-config-konnektid/rules/babel",
+        "eslint-config-konnektid/specs/module",
         "eslint-config-konnektid/rules/module"
     ]
 };