+++ react.js
@@ -1,7 +1,7 @@
 module.exports = {
     extends: [
-        "eslint-config-konnektid",
-        "eslint-config-konnektid/rules/babel",
-        "eslint-config-konnektid/rules/jsx",
+        "eslint-config-konnektid/babel",
+        "eslint-config-konnektid/specs/jsx",
+        "eslint-config-konnektid/rules/react"
     ]
 };