+++ rules/base/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+    extends: [
+        "eslint-config-konnektid/rules/base/preventability",
+        "eslint-config-konnektid/rules/base/bestpractices",
+        "eslint-config-konnektid/rules/base/variability",
+        "eslint-config-konnektid/rules/base/brevity",
+        "eslint-config-konnektid/rules/base/breathability",
+        "eslint-config-konnektid/rules/base/maintainability",
+        "eslint-config-konnektid/rules/base/snazzy",
+    ]
+};