+++ rules/module.js
@@ -1,5 +1,7 @@
 module.exports = {
-    parserOptions: {
-        sourceType: "module"
+    rules: {
+
+        // ignore extra strict directives
+        "strict": 0
     }
-}
+};