+++ universal.js
@@ -1,12 +1,10 @@
 module.exports = {
     extends: [
         "eslint-config-konnektid/specs/es7",
-        "eslint-config-konnektid/specs/jsx",
-        "eslint-config-konnektid/envs/browser",
+        "eslint-config-konnektid/envs/universal",
         "eslint-config-konnektid/envs/webpack",
         "eslint-config-konnektid/rules/base",
         "eslint-config-konnektid/rules/webpack",
-        "eslint-config-konnektid/rules/frontend",
-        "eslint-config-konnektid/rules/react"
+        "eslint-config-konnektid/rules/babel"
     ]
 };